#13 Output results. #12 calculate k-l divergence #8 update to tools #6 added meters

+++ src/protocol_model/src/DirectionChangesMeter.cpp
@@ -31,5 +31,5 @@ void DirectionChangesMeter::calculateMeasurement(Flow* flow,
         count++;
     }
 
-    _fingerprint[0] = count / flow->pktCount();
+    _fingerprint[0] = (double)count / flow->pktCount();
 }
